Fix typos and ensured example works
