--- example/.~lock.example.csv#
@@ -1,1 +1,0 @@
-,stephen.crowell,banrock-7163,20.10.2022 12:34,file:///home/local/KHQ/stephen.crowell/.config/libreoffice/4;
